extract game specific code to separate file
